add food tag to GetFoodQuery

+++ src/Infrastructure/Contracts/AzureTables/FoodTag.cs
@@ -5,8 +5,8 @@
         public string Id { get; set; }
         public string Name { get; set; }
         public string Mark { get; set; }
-        public float MinDailyValuePercentage { get; set; } = 0;
-        public float MaxDailyValuePercentage { get; set; } = float.MaxValue;
+        public double MinDailyValuePercentage { get; set; } = 0;
+        public double MaxDailyValuePercentage { get; set; } = int.MaxValue;
         public string NutrientId { get; set; }
         public string NutrientName { get; set; }
     }
